rest api

+++ HRTool/src/main/java/app/controllers/ProjectRestController.java
@@ -2,14 +2,19 @@ package app.controllers;
 
 
 import java.util.List;
+import java.util.Set;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.rest.core.annotation.RepositoryEventHandler;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
 
+import app.model.Employee;
 import app.model.Project;
 import app.model.ProjectInfo;
 import app.repository.ProjectInfoRepository;
@@ -17,7 +22,7 @@ import app.repository.ProjectRepository;
 
 
 @RestController
-@RequestMapping("projects")
+@RequestMapping("/resources/projects")
 @RepositoryEventHandler(Project.class)
 public class ProjectRestController {
 
@@ -27,19 +32,57 @@ public class ProjectRestController {
 	@Autowired
 	private ProjectInfoRepository projectInfoRepository;
 	
+	
+	/**
+	 * 
+	 * @param {Project} id
+	 * @return {ResponseEntity} of {Project} infoes for given {Project} id
+	 *
+	 */
+	@RequestMapping(value="/{id}/projectInfoes", method = RequestMethod.GET)
+	public ResponseEntity getProjectInfoes(@PathVariable("id") int id){
+		
+		return new ResponseEntity(projectRepository.findOne(id).getProjectInfo(), HttpStatus.OK);
+	}
+	
+	
+	/**
+	 * 
+	 * @param emp - {Project} for save
+	 * @return {ResponseEntity} with {HttpStatus.OK}
+	 */
+	@RequestMapping(method = RequestMethod.PUT)
+	public ResponseEntity put(@RequestBody Project proj){
+		
+		projectRepository.save(proj);
+		return new ResponseEntity(HttpStatus.OK);
+	}
+
+	
+	/**
+	 * Deletes project with given id
+	 * @param id - {Project} id for deletion
+	 * @return {HttpStatus.NOT_FOUND} or {HttpStatus.OK}
+	 */
 	@RequestMapping(value="/{id}", method = RequestMethod.DELETE)
-	public Project deleteProject(@PathVariable("id") int id){
+	public ResponseEntity deleteProject(@PathVariable("id") int id){
+		
+		if(projectRepository.findOne(id) == null) 
+			return new ResponseEntity(HttpStatus.NOT_FOUND);
+
 		List<ProjectInfo> projectInfoes = projectInfoRepository.findAll();
 		for (ProjectInfo pi : projectInfoes){
 			if (pi.getProject().getIdProject() == id){
-			//	pi.setEmployee(null);
-			//	pi.setProject(null);
 				pi.setTagClouds(null);
 				projectInfoRepository.delete(pi.getIdProjectInfo());
 			}
 		}
+		
 		projectRepository.delete(id);
- 		return null;
+		return new ResponseEntity(HttpStatus.OK);
 	}
+
+	
+
 	
 }
